web-app/requirements.txt
Flask==2.2.3
openai==0.27.4
pinecone-client==2.2.1
gunicorn==20.1.0
firebase-admin
